Add spaced repetition scheduling
Adds mechanism to keep track on user's progress
on each card, and to reschedule it after each repetition
to a specific point of time in future, based on previous
intervals and user's answer.

The initial implementation uses pows of 2
as intervals:
-1 0 1 2 4 8 16 ...

Currently a "day" is 24h, this approach should
be refined. Minute-to-minute granularity brings
nothing except partial assignments which are
an unnecessary complication leading to irritation
and bad user experience.

A debug view was added to keep a visual track on
a card's progress. During learning, it's shown
in the bottom part of the screen. It shows up only
for debug builds and can be turned off in code.