Clever adding - add forward and reversed cards
When a card is added, a respective reversed
card will be added as well, one per each
translation the original card has.

Added LanguagesRegistry as a temporary hack
to have the original and the translations
languages somewhere. In future, this information
should come from the Domain object.

Cards have a field CardType to distinguish
whether it's a forward or a reverse card.