Allow deleting a card
Add a menu item on the learning flow,
which allows to remove the current card
with a confirmation dialog.

Deleting a card won't relete a respective
reverse card.

Deleting a card will also delete all the
expressions which are no more used in any
of the existing cards.

cardById() was brought back for the sake
of testing.