Add bottom navigation
Google design support lib was rejected in favor of a custom library,
because design support widget enforces ugly UI for 4 tabs case
with no means to reset it (other that reflection which I won't use).
It was added as a dependency to build.gradle though, because it's
required by the custom lib.