Support domain model for languages
Domain represents a pair of two languages -
one for forward and one for reverse sides
of cards. Logically it corresponds to one
foreign language that the user is studying.

Each card and deck belong to one and only
one domain. Languages and expressions are
independent of domains.

Currently one default domain will be created
on each app start.