[Settings] Add learning a specified amount of cards
In the settings fragment, use can specify the limits
of new and review cards to learn per day. There is no
limit for the relearn cards, all of them are treated
as pending.

On the first app start, the default values will be
written to the preferences. For now it's 15 for daily
new cards and 30 for daily review cards.

Journal was added to store the information about how
many cards were learned on a specific date. It is
used by LearningFlow when creating a pending
assignment, and may be used in future for creating
statistics.

FirstStart was added to do initialization on the
first app start.

For the settings, support preference library was
denied in favor of a custom library
com.takisoft.fix:preference-v7 which is basically
a wrapper over the support lib with some fixes and
improvements. Reasons:
1. support lib version is full of bugs
2. extended functionality of takisoft lib
   (e.g. AutoSummaryEditTextPreference)

SettingsFragment will use not a shared preference
instance but a custom PreferenceDataStore which
reads/writes data from/to the app's Preferences.
This was done for the sake of keeping code clean
and modularized.