Allow deleting an empty deck
For non-empty deck, operation will be rejected.

Storing of default deck id was removed with this
feature, as I don't want to handle cases like -
what to count as default deck if user removes
the default deck. From now on, if user has no decks,
a default deck will be created on new app start.