[Backup] Add ability to create backups
Backup-er will use a special interface to
the repostory which allows more direct access
to data and pagination. It has no visible
connection to Repository interface itself,
though logically it is the same database.

Data will be exported to a hardcoded directory
sdcard/Coriolan/backup
Each file will have .coriolan extention.
Backup format is simple JSON.

I use Jackson to deal with JSON on a low level,
as for now it seems to be the only elegant
way to work with Kotlin (allowing usage of
data classes, default parameners etc.).
Maybe it can be revisited in future to use
benefits of object-to-json conversion.
JSON is parsed on fly to support big databases.