[Design] Give visual feedback on the last answer
A new TouchInterceptorView was added to handle it.
It accepts one or multiple anchors and catches all
touch events addressed to them, applies to itself
and then passes to the respective anchor.
This way we can create an overflow ripple to fill
the whole screen with a color when user taps a button.