Get rid of TodayProvider interface
The logic of getting the "today" date, and the date itself, is a part of the most global state for the application. It can be requested from any point of the app. There is no need to substitute or mock it. Thus, there is no need to make it an injectable interface. Static access is much more suitable for this case.