Improve SQLite columns naming
Due to limitations of Android cursor implementation, we cannot refer to columns in "Table.Column" format in cursor output. Thus, I am renaming all columns to contain the name of the table, so that there are no conflicting names. This makes queries generally more lightweight, and allows to do more in queries.